+++ lib/include/resdata/well/well_branch_collection.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstddef>
 #include <memory>
 
 #include <ert/util/type_macros.hpp>
@@ -11,8 +12,8 @@ well_branch_collection_type *well_branch_collection_alloc();
 void well_branch_collection_free(well_branch_collection_type *branches);
 bool well_branch_collection_has_branch(
     const well_branch_collection_type *branches, int branch_id);
-int well_branch_collection_get_size(
-    const well_branch_collection_type *branches);
+size_t
+well_branch_collection_get_size(const well_branch_collection_type *branches);
 const std::shared_ptr<WellSegment> well_branch_collection_get_start_segment(
     const well_branch_collection_type *branches, int branch_id);
 bool well_branch_collection_add_start_segment(
