+++ lib/resdata/smspec_node.cpp
@@ -845,42 +845,42 @@ rd_smspec_var_type smspec_node::valid_type(const char *keyword,
    rd_smspec_fread_header() functions in addition. UGGGLY
 */
 
-smspec_node::smspec_node(int param_index, const char *keyword, int num,
+smspec_node::smspec_node(size_t param_index, const char *keyword, int num,
                          const char *unit, const int grid_dims[3],
                          float default_value, const char *key_join_string)
     : smspec_node(param_index, keyword, nullptr, num, unit, grid_dims,
                   default_value, key_join_string) {}
 
-smspec_node::smspec_node(int param_index, const char *keyword, int num,
+smspec_node::smspec_node(size_t param_index, const char *keyword, int num,
                          const char *unit, float default_value,
                          const char *key_join_string)
     : smspec_node(param_index, keyword, nullptr, num, unit, nullptr,
                   default_value, key_join_string) {}
 
-smspec_node::smspec_node(int param_index, const char *keyword,
+smspec_node::smspec_node(size_t param_index, const char *keyword,
                          const char *wgname, const char *unit,
                          float default_value, const char *key_join_string)
     : smspec_node(param_index, keyword, wgname, nums_unused, unit, nullptr,
                   default_value, key_join_string) {}
 
-smspec_node::smspec_node(int param_index, const char *keyword,
+smspec_node::smspec_node(size_t param_index, const char *keyword,
                          const char *wgname, int num, const char *unit,
                          float default_value, const char *key_join_string)
     : smspec_node(param_index, keyword, wgname, num, unit, nullptr,
                   default_value, key_join_string) {}
 
-smspec_node::smspec_node(int param_index, const char *keyword, const char *unit,
-                         float default_value)
+smspec_node::smspec_node(size_t param_index, const char *keyword,
+                         const char *unit, float default_value)
     : smspec_node(param_index, keyword, nullptr, nums_unused, unit, nullptr,
                   default_value, nullptr) {}
 
 //copy constructor with a new id
-smspec_node::smspec_node(const smspec_node &node, int param_index) {
+smspec_node::smspec_node(const smspec_node &node, size_t param_index) {
     *this = node;
     this->params_index = param_index;
 }
 
-smspec_node::smspec_node(int param_index, const char *keyword,
+smspec_node::smspec_node(size_t param_index, const char *keyword,
                          const char *wgname, int num, const char *unit,
                          const int grid_dims[3], float default_value,
                          const char *key_join_string) {
@@ -991,7 +991,7 @@ smspec_node::smspec_node(int param_index, const char *keyword,
     set_gen_keys(key_join_string);
 }
 
-smspec_node::smspec_node(int param_index_, const char *keyword_,
+smspec_node::smspec_node(size_t param_index_, const char *keyword_,
                          const char *wgname_, const char *unit_,
                          const char *lgr_, int lgr_i, int lgr_j, int lgr_k,
                          float default_value_, const char *key_join_string_) {
