+++ lib/resdata/tests/rd_kw_space_pad.cpp
@@ -27,7 +27,7 @@ int main(int argc, char **argv) {
         FILE *stream = util_fopen("file", "r");
         char buffer[8];
         size_t offset = 4 + 16 + 4 + 4;
-        fseek(stream, offset, SEEK_SET);
+        fseek(stream, static_cast<long>(offset), SEEK_SET);
         fread(buffer, 1, 8, stream);
         for (int i = 0; i < 8; i++)
             test_assert_int_equal(buffer[i], ' ');
